322 lab 1

+++ comp322/labs/assignment1.c
@@ -8,16 +8,17 @@ each choice calls the appropriate procedure, where the choices are:
 3) Destroy all descendants of a process
 4) Quit program and free memory */
 
+/***************************************************************/
+
 /* Define structures and global variables */
 
-// TODO: main
-
 int n;
+int initialized = 0;
 
 struct node {
     int process_id;
     struct node* link;
-}
+};
 
 typedef struct node list_type;
 
@@ -31,35 +32,41 @@ typedef struct pcb pcb_type;
 
 /***************************************************************/
 
+int check_init() {
+
+    if (initialized == 0) {
+        printf("Program not initialized - please choose option 1 on main menu\n");
+        return 0;
+    } else return 1;
+
+}
+
 void print_pcbs() {
+
     /* declare local vars */
     int i;
     list_type* child;
 
     /* for each PCB index i from 0 up to (but not including) maximum number*/
     for (i = 0; i < n; i++) {
-
         /* check if PCB[i] has a parent and children */
-        if ( (pcb_array[i].parent > -1) && (pcb_array[i] != NULL) ) {
+        if ( (pcb_array[i].parent > -1) && (&pcb_array[i] != NULL) ) {
             /* print message about children of the current PCB[i] */
-
             /* initialize variable to head of list of children */
             child = pcb_array[i].children;
-
+            printf("Process %d\n", i);
+            printf("--Children:\n");
             /* while the end of the linked list of children is not reached */ 
             while (child != NULL) { 
                 /* print message about current child process index */
-                printf("Process %d is the parent of Process %d\n", pcb_array[i].parent, child->process_id);
-                
+                printf("\tProcess %d", child->process_id);
                 /* move to next node in linked list */
                 child = child->link;
             } /* while */
 
             /* print newline */
             printf("\n");
-
         }/* if */
-
     } /* for */
 
     return;
@@ -69,6 +76,7 @@ void print_pcbs() {
 /***************************************************************/
 
 void input_params() {
+
     /* declare local vars */
     int i;
 
@@ -82,18 +90,27 @@ void input_params() {
     /* Define PCB[0] */
     pcb_array[0].parent = 0;
     pcb_array[0].children = NULL;
+    initialized = 1;
 
     /* for-loop to intitialize all other indices' parent to -1 */
     for (i = 1; i < n; i++) {
         pcb_array[i].parent = -1;
         pcb_array[i].children = NULL;
     }
+
+    printf("New 1st ancestor created.\n");
+    printf("Memory remaining: %d\n", n - 1);
+
     return;
+
 } /* end of procedure */
 
 /***************************************************************/
 
 void create_child() {
+
+    if (check_init() == 0) return; 
+
     /* define local vars */
     int p;
     int i = 1;
@@ -106,7 +123,7 @@ void create_child() {
 
     /* search for first available index q without a parent in a while 
     loop */
-    while (i < n && pcb_array[i].parent != -1) {
+    while ( (i < n) && (pcb_array[i].parent != -1) ) {
         i++;
     }
     if (i == n) {
@@ -130,12 +147,14 @@ void create_child() {
     } else {
         curr = pcb_array[p].children;
 
-        while(curr.link != NULL) {
-            curr = curr.link;
+        while(curr->link != NULL) {
+            curr = curr->link;
         }
-        curr.link = child;
+        curr->link = child;
     }
     /* call procedure to print current hierachy of processes */
+    printf("New process created: Process %d\n", i);
+    printf("Memory remaining: %d\n", n - i - 1);
     print_pcbs();
     return;
 } /* end of procedure */
@@ -143,6 +162,7 @@ void create_child() {
 /***************************************************************/
 
 void kill_process(list_type* node) {
+
     /* declare local vars */
     int q;
 
@@ -155,9 +175,10 @@ void kill_process(list_type* node) {
 
         /* set variable q to current node's process index field */
         q = node->process_id;
-
+        
         /* call self on children of PCB[q] */
         kill_process(pcb_array[q].children);
+        printf("Destroying process %d\n", q);
 
         /* free memory of parameter */
         free(node);
@@ -167,7 +188,6 @@ void kill_process(list_type* node) {
 
         /* set parameter to NULL */
         node = NULL;
-
     } /* end of else */
 
     return;
@@ -177,12 +197,20 @@ void kill_process(list_type* node) {
 /***************************************************************/
 
 void destroy_descendants() {
+
+    if (check_init() == 0) return; 
+    
     /* declare local vars */
     int p;
 
     /* prompt for parent process index p */
     printf("Enter the parent process ID: ");
     scanf("%d", &p);
+
+    if (p < 0 || p > n) {
+        printf("Invalid process ID!");
+        return;
+    }
 
     /* call recursive procedure to destroy children of PCB[p] */
     kill_process(pcb_array[p].children);
@@ -193,15 +221,17 @@ void destroy_descendants() {
     /* call procedure to print current hierarchy of processes */
     print_pcbs();
 
+    printf("Finished destroying processes.\n");
     return;
+
 } /* end of procedure */
 
 /***************************************************************/
 
 void exit_program() {
+
     /* check if PCB is non null) */
     if (pcb_array == NULL) {
-        printf("Error: PCB array doesn't exist.\n");
         return;
     } else {
         /* check if children of PCB[0] is not null */
@@ -213,28 +243,35 @@ void exit_program() {
     } /* if */
 
     return;
+
 } /* end of procedure */
 
 /***************************************************************/
 
 int main() {
+
     /* declare local vars */
     int input = 0;
 
     /* while user has not chosen to quit */
     while (input != 4) {
         /* print menu of options */
-        printf();
+        printf("\nPlease choose from the following:\n");
+        printf("1) Initialize process tree\n");
+        printf("2) Create a new child process\n");
+        printf("3) Destroy all descendants of a process\n");
+        printf("4) Quit program and free memory\n");
         /* prompt for menu selection */
         printf("Enter selection: ");
         scanf("%d", &input);
+        printf("\n");
         /* call appropriate procedure based on choice--use switch 
         statement or series of if, else if, else statements */
         switch (input) {
-            case 1:
-            case 2:
-            case 3:
-            case 4:
+            case 1: input_params(); break;
+            case 2: create_child(); break;
+            case 3: destroy_descendants(); break;
+            case 4: exit_program(); break;
             default: printf("Invalid entry, please try again...");
         }
     } /* while loop */
